% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNLL.R
\name{CoNLL}
\alias{CoNLL}
\title{CoNLL class.}
\description{
CoNLL class.

CoNLL class.
}
\details{
Toolset for a workflow to reshape a CoNLL input to a CWB compatible format.
}
\examples{
if (FALSE){
  use("GermaParl")
  filename <- system.file(package = "polmineR.anno", "extdata", "CoNLL", "16_126_FDP_Markus_Loening_2871305.tsv")
  P <- partition("GERMAPARL", text_lp = "16", text_protocol_no = "126", text_speaker = "Markus Loening")
  Anno <- CoNLL$new(filename = filename, partition = P)
  Anno$getCorpusPositions()
  annotations <- Anno$cpos
  cpos <- unlist(apply(annotations[id == "B-15_15"], 1, function(row) row["cpos_left"]:row["cpos_right"]))
  html(P) \%>\% highlight(highlight = list(yellow = cpos)) \%>\% htmltools::html_print()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tab}}{STORY TO BE TOLD}

\item{\code{filename}}{character string, the filename of the CoNLL annotation}

\item{\code{partition}}{a partition object, generated on the basis of sAttributes}

\item{\code{cpos}}{a data.table with the columns "cpos_left" and "cpos_right" defining the regions of the annotations}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CoNLL$new()}}
\item \href{#method-getCorpusPositions}{\code{CoNLL$getCorpusPositions()}}
\item \href{#method-clone}{\code{CoNLL$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoNLL$new(filename, partition)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{name of file to be processed}

\item{\code{partition}}{partition where to locate annotations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCorpusPositions"></a>}}
\if{latex}{\out{\hypertarget{method-getCorpusPositions}{}}}
\subsection{Method \code{getCorpusPositions()}}{
STORY TO BE TOLD
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoNLL$getCorpusPositions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoNLL$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
