% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxqda-package.R
\docType{package}
\name{maxqda-package}
\alias{maxqda-package}
\alias{maxqda}
\title{maxqda-package}
\description{
Process the MAXQDA database.
}
\references{
http://polmine.sowi.uni-due.de
}
\author{
Andreas Blaette (andreas.blaette@uni-due.de)
}
\keyword{package}
