% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAXQDA.R
\name{MAXQDA}
\alias{MAXQDA}
\title{MAXQDA.}
\description{
MAXQDA.

MAXQDA.
}
\details{
Class to manage and get annotations from a MAXQDA file that stores annotations
to html documents that have previously been generated by polmineR. These html
documents are required to have character offset positions in tags.
}
\examples{
maxfile <- system.file(package = "maxqda", "extdata", "maxqda", "reuters.mx12")
htmlDir <- system.file(package = "maxqda", "extdata", "html", "reuters")
M <- MAXQDA$new(dbFilename = maxfile, htmlDir = htmlDir)
annotationsDT <- M$getAll()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dbFilename}}{the MAXQDA file (*.mx12) from which annotations are to
be extracted (full path)}

\item{\code{codings}}{table with codings/annotations derived from the MAXQDA
file, a \code{data.table}}

\item{\code{codewords}}{table with codewords derived from the MAXQDA file, a
\code{data.table}}

\item{\code{texts}}{table with texts derived from the MAXQDA file, a
\code{data.table}}

\item{\code{htmlDir}}{directory where the html files that have been imported
into MAXQDA reside}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MAXQDA$new()}}
\item \href{#method-getMaxqdaData}{\code{MAXQDA$getMaxqdaData()}}
\item \href{#method-getOffsetDT}{\code{MAXQDA$getOffsetDT()}}
\item \href{#method-getOne}{\code{MAXQDA$getOne()}}
\item \href{#method-getAll}{\code{MAXQDA$getAll()}}
\item \href{#method-clone}{\code{MAXQDA$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of class \code{MAXQDA}; the param \code{dbFilename}
indicates the MAXQDA database, the \code{htmlDir} the directory where
html files as generated with \code{polmineR} reside
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAXQDA$new(dbFilename, htmlDir)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dbFilename}}{filename of the MAXQDA database (file ending: ".mx12")}

\item{\code{htmlDir}}{html directory}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMaxqdaData"></a>}}
\if{latex}{\out{\hypertarget{method-getMaxqdaData}{}}}
\subsection{Method \code{getMaxqdaData()}}{
Fill the fields \code{codings}, \code{codewords}, and \code{texts} with
\code{data.table} by extracting the respective tables from the MAXQDA
file specified by field \code{dbFilename}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAXQDA$getMaxqdaData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getOffsetDT"></a>}}
\if{latex}{\out{\hypertarget{method-getOffsetDT}{}}}
\subsection{Method \code{getOffsetDT()}}{
Extract \code{data.table} with character offset positions and corpus
positions from the html file specified by param \code{filename}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAXQDA$getOffsetDT(filename)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{filename of a html file that has been imported into MAXQDA}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getOne"></a>}}
\if{latex}{\out{\hypertarget{method-getOne}{}}}
\subsection{Method \code{getOne()}}{
Extract annotations for one document; essentially a worker called by the
method \code{getAll}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAXQDA$getOne(textId, offsetDT, corpus = "REUTERS")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{textId}}{DONTKNOW}

\item{\code{offsetDT}}{DONTKNOW}

\item{\code{corpus}}{corpus from which the partition is derived}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getAll"></a>}}
\if{latex}{\out{\hypertarget{method-getAll}{}}}
\subsection{Method \code{getAll()}}{
Extract annotations for all documents that reside in the \code{htmlDir}
that has been specified upon initialization
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAXQDA$getAll(verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{logical, whether to be talkative}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAXQDA$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
