% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polmineR.anno_package.R
\docType{package}
\name{polmineR.anno-package}
\alias{polmineR.anno-package}
\title{polmineR.anno-package}
\description{
polmineR.anno-package
}
\references{
http://polmine.sowi.uni-due.de
}
\author{
Andreas Blaette (andreas.blaette@uni-due.de)
}
\keyword{package}
