% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNLL.R
\docType{data}
\name{CoNLL}
\alias{CoNLL}
\title{CoNLL class.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CoNLL
}
\arguments{
\item{filename}{name of file to be processed}

\item{partition}{partition where to locate annotations}
}
\description{
Toolset for a workflow to reshape a CoNLL input to a CWB compatible format.
}
\section{Fields}{

\describe{
\item{\code{filename}}{character string, the filename of the CoNLL annotation}

\item{\code{cpos}}{a data.table with the columns "cpos_left" and "cpos_right" defining the regions of the annotations}

\item{\code{partition}}{a partition object, generated on the basis of sAttributes}
}}

\section{Methods}{

\describe{
  \item{\code{$new(filename, partition)}}{}
  \item{\code{$load()}}{Read the content of the file specified upon initialization by \code{filename},
  and keep the parsed input in the slot \code{tab}}
  \item{\code{$getCorpusPositions}}{}
}
}

\examples{
if (require("GermaParl")){
  use("GermaParl")
  filename <- system.file(package = "polmineR.anno", "extdata", "CoNLL", "16_126_FDP_Markus_Löning_2871305.tsv")
  P <- partition("GERMAPARL", text_lp = "16", text_protocol_no = "126", text_speaker = "Markus Löning")
  Anno <- CoNLL$new(filename = filename, partition = P)
  Anno$getCorpusPositions()
  annotations <- Anno$cpos
  cpos <- unlist(apply(annotations[id == "B-15_15"], 1, function(row) row["cpos_left"]:row["cpos_right"]))
  html(P) \%>\% highlight(highlight = list(yellow = cpos)) \%>\% htmltools::html_print()
}
}
\keyword{datasets}
