% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxqda-package.R
\docType{package}
\name{maxqda-package}
\alias{maxqda-package}
\alias{_PACKAGE}
\alias{maxqda}
\title{maxqda: Use MAXQDA for Corpus Annotations}
\description{
To learn more about the 'maxqda' package,
start with the vignette: \code{browseVignettes(package = "maxqda")}
}
\references{
Kuckartz, Udo; Raediker, Stefan (2019): \emph{Analyzing Qualitative Data
with MAXQDA.} Wiesbaden: Springer.
}
\author{
Andreas Blaette (andreas.blaette@uni-due.de)
}
\keyword{package}
