% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAXQDA.R
\docType{data}
\name{MAXQDA}
\alias{MAXQDA}
\title{MAXQDA.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MAXQDA
}
\description{
Class to manage and get annotations from a MAXQDA file that stores annotations
to html documents that have previously been generated by polmineR. These html
documents are required to have character offset positions in tags.
}
\section{Fields}{

\describe{
\item{\code{dbFilename}}{the MAXQDA file (*.mx12) from which annotations are to be extracted (full path)}

\item{\code{codings}}{table with codings/annotations derived from the MAXQDA file, a \code{data.table}}

\item{\code{codewords}}{table with codewords derived from the MAXQDA file, a \code{data.table}}

\item{\code{texts}}{table with texts derived from the MAXQDA file, a \code{data.table}}

\item{\code{htmlDir}}{directory where the html files that have been imported into MAXQDA reside

#' @section Arguments:
\describe{
  \item{dbFilename}{filename of the MAXQDA database (file ending: ".mx12")}
  \item{htmlDoc}{html document as generated by \code{html} method (required to include offset)}
  \item{corpus}{corpus from which the partition is derived}
  \item{filename}{filename of a html file that has been imported into MAXQDA}
  \item{verbose}{logical, whether to be talkative}
}}
}}

\section{Methods}{

\describe{
  \item{\code{initialize((dbFilename, htmlDir))}}{Create a new instance of 
  class \code{MAXQDA}; the param \code{dbFilename} indicates the MAXQDA 
  database, the \code{htmlDir} the directory where html files as generated 
  with \code{polmineR} reside} \item{\code{getMaxqdaData()}}{Fill the fields
  \code{codings}, \code{codewords}, and \code{texts} with \code{data.table}
  by extracting the respective tables from the MAXQDA file specified by field
  \code{dbFilename}} \item{\code{getOffsetDT(filename)}}{Extract
  \code{data.table} with character offset positions and corpus positions from
  the html file specified by param \code{filename}} 
  \item{\code{getOne(textId, offsetDT, corpus = "REUTERS")}}{Extract
  annotations for one document; essentially a worker called by the method
  \code{getAll}} \item{\code{getAll(verbose = TRUE)}}{Extract annotations for
  all documents that reside in the \code{htmlDir} that has been specified
  upon initialization}
}
}

\examples{
maxfile <- system.file(package = "polmineR.anno", "extdata", "maxqda", "reuters.mx12")
htmlDir <- system.file(package = "polmineR.anno", "extdata", "html")
M <- MAXQDA$new(dbFilename = maxfile, htmlDir = htmlDir)
annotationsDT <- M$getAll()
}
\keyword{datasets}
